Choose rule algorithm change
This will enable rule choosing match Rule Objects directly rather than
doing check on whether matche's root link is a BindLink or not. Also
circumvents the issue of Node var matching which was overlooked to be
fixed in the last commit.

Working directly with Rule objects avoids unnecessary dependency on Rule's
structural representation in the atomspace.