Fix PatternMatchEngine::variable_compare for unscoped variables
That is if a bound variable used in the pattern matcher is reused in a
grounded tree in a scoped manner then allow grounding.

Add unit test that illustrate that see TypeChoiceUTest::test_connected.

The pattern is

GetLink
   X Y
   ImplicationLink
      X
      Y

and the grounded term is

ImplicationLink
   PredicateNode "P"
   LambdaLink
      X
      EvaluationLink
         PredicateNode "P"
         X

The same X is used in both the pattern and the grounded term, but in the
grouned term it is scoped, so it's OK.