Implement MergeCtrl and pass it to TruthValue::merge
The structure allows to control how to merge truth values. For the
moment there are only a few options, such as how to the merge the tv
type, older, newer, stronger or weaker as previously implemented and
described in the issue https://github.com/opencog/opencog/issues/1295

In additional it allows to control what formula to use, for now only 2
formulae are implemented. The revision rule as described in the PLN
book, and a simpler retain-tv-with-higher-confidence rule.