Replace pair <forward, backward> by list <forward, [backward-1], ...>
We possibly need multiple backward rules because a single forward rule
can output different patterns (think of equivalent to double implication
A<->B, A->B, B->A). Ultimately this could be replaced by a single
backward rule if the pattern matcher were supporting disjunctions, but
even though it's still more convenient to have such an expansion.