Take care of insterting in _bit_as all bit parts
This is to workaround the fact that alpha_equivalent scope links will
point to the same scope links when it is in a atomspace. This is not
very convenient for building forward chaining strategies, but
ultimately fcs expansion should be fixed so that it doesn't need to
work around that on-the-fly alpha-conversion feature.