Hacky fix for #1739
Bug #1739 reports some "obvious" hash collisions. This is a dopey fix.
This is still a collision-prone hash but maybe more tolerable.
I'm trying to avoid fnv1a because I figure that hash computation
is a significant part of the performance of atom creation, and I don't
want to slow that down.