Take into account quote link when recording subclause pattern
Take into account quote link when recording association between
subclause pattern and its grounding, which result into fixing the bug
exhibited by GetLinkUTest::test_quote_equal.

Note that it only works because the pattern in argument of EqualLink
is equal to the subpattern of a real clause. In the cases where the
pattern of the virtual clause would differ from any pattern in real
clause, this trick would no longer and proper quotation consumption
would need to be carried out. However this probably cover most
practical cases, and taking quote link in the association bettween
subclause pattern and grounding should be done nonetheless.