Add green ball unit test for the URE
Inspired from Alexander Scherbatiy network flow problem. A piece of
evidence is required to be built (such that a green is ball given that
it only absorbs red and blue) which requires the direct evaluation
rule to keep track of each piece of evidence in order to back-infer
the missing piece of evidence.

It requires GlobNode to be supported (or at least it would simplify
the rule) so I'm postponing its completion till GlobNode is supported.