Make Variables::_ordered = false the default
Only care about order of VariableList is used. In all other
cases (such as empty or single variable) then _ordered is set to
false, that is order does not matter.

As a result the following scopes are equivalent

Scope
  <body>

Scope
  Variable "$X"
  <body>

Scope
  VariableSet
    Variable "$X"
  <body>

assuming <body> only contains $X as free variable.

Thus,

Scope
  Variable "$X"
  <body>

is NOT equivalent to

Scope
  VariableList
    Variable "$X"
  <body>

due to the explicit use of VariableList. Maybe it should be, that is
to be determined.