Assume FreeLinks are always ordered links.
This reverts commit e3c26feac825de54db8461484df446583a49dd77
which added this code, with minimal explanation as to what the
actual problem was.

The current problem is that the C++ code comment mentions
EqualLink and IdenticalLink, both of which *are* unordered,
but neither of which are FreeLinks.  Grepping through the
atom type definitions, I found no free, unordered link types.
(for the spacetime, attention and agi-bio).