Fix get-snail in VirtualUTest
Indeed now that FreeVariables::substitute_scoped is fixed, in the
absence of type restriction for variable x and y in get-snail, y
becomes a valid candidate since it inherits animal, there exists
another y (locally z), such that y inherits from.

In order to fix the get-snail query we restrict x and y to be concept
nodes.