Fix potential corruption noted in pull req #2505
In the pull-req comments for #2505,
This risks a memory-access fault, or other corruption, because
hidden_map goes out of scope, and it's destructor runs, with
the closing brace below. I think it's some happy accident that
it didn't die ... probably because its on the stack, and the dtor
won't run until the stack is popped.