Restore stubbed-out unit test.
Placing a check in our code base because some other unknown
third-party code might be broken according to a post on the
internet does not make sense in any rational way that I can
understand it.  If other users have bugs in their code, that
should not be our problem, nor is it our duty to perform a
check on their behalf.