Relax boolean link validator
Currently the validator associated to BOOLEAN_LINK is
check_evaluatable, the problem is boolean links can also be used on
atoms that are not evaluatable, such as implication links, according
to PLN.

These links cannot be made evaluatable without invalidating the
function is_constant in PatternUtils.h.  So I think the best thing for
now is to relax the validator of BOOLEAN_LINK.  I have relaxed
check_evaluatable, however I suppose the latter should either be
renamed check_boolean, or a new check_boolean validator should be
introduced that uses check_evaluatable + the additional code in this
commit.