Fix truly annoying issue with TruthValueKey
This is possibly just papering over which might be an old design
flaw. The problem is that TruthValues are stored with a special
PredicateNode "*-TruthValueKey-*" which is not normally in the
AtomSpace.  However, file reads and network transfers can result
in it showing up there, and then truth-value lookup fails, because
the wrong key is used. So this works around that...

It's not a good solution, but the other one, of forcing it into
the atomspace, cause more tha 40 unit tests to fail. I tried to
bite the bullet on that one, but the pain was too great.  So
instead we have this fix.