Revert "Merge pull request #2500 from linas/anchor-link"
Linking results to the AnchorNode idea now seems like a bad idea,
in retrospect.  The ideas in #2911 seem superior. So trash the
AnchorNode support in the query subsystem.

FWIW, if this kind of thing was wanted, a better solution would
be a new kind of Value, that dequeued from a QueueValue, and
plopped the results onto an AnchorNode. ***This*** is the real
reason for stripping away this code: its not generic enough.

This reverts commit 46dea8ea1fe8f81c079bc770861d59aad448f2bb, reversing
changes made to 5ec84bbe77daa06af72e39bb5b2e04e72c7ddcf8.