Buffers less than 1000 are always drained.
Intended use is to buffer millions, not dozens. Some of the important
atoms are few in number, and are heavily deduplicated; we want to
write all of them, instead of being hostage to the whims of std::set
ordering of handles. (Which prevents the far end of the set from getting
flushed)