Fix screwy CondLink behavior.
The old instantiator just does the wrong thing for evaluatables.
And if the user really wanted to evaluate teh evaluatable result,
well, that is something they need to do on their own, instead of
expecting magical evades-the-rules results.

This is just prehistoric design cruft continuing to take a toll
on the debuggability of the code base.