Only exposed ValueShims upset the applecart
ValueShims inside executale terms don't count, because execution
will typically wipe them out. Check this case before execution.

An alternative implementation that passes unit test would be
to try-catch the execution, and catch the attempted insertion
of ValueShims into the AtomSpace. But this is dangerous, because
the execution could have side effects that we don't control for.