Disable guile/scheme handles to orphan atoms
Allowing guile/scheme to work with atoms that are not in any
AtomSpace causes extensive confusion and weird bugs. Atoms may
end up homeless because they have been extracted from the current
AtomSpace, or because the AtomSpace that contained them was deleted.
In either case, guile can retain handles to that/those Atoms. Those
handles should be consistently treated as invalid, by all of the
guile API, and not just some of it. This fixes that.