Python: remove finalize()
It makes no sense to tel python to self-destruct. I think the initial
idea was to stop opencog, rathre than stoping python, but scope creep
beeing what it is ... anyway, I don't know how to "stop opencog" or
who would want to do that from python, or what the knock-on effects
would be since the atomspace is a mutli-user system (e.g. in the
cogserver -- if someone called "stop opencog" inside the cogserver,
things would get ugly. But stopping python inside the cogserver does
make sense ... but all that would need careful rethinking, which is
not on the menu this evening.)