Fighting off yet moe URE brain-damage
No wonder URE never worked right. It's an endless litany of
depending on ill-defined or illegal behavior, and then friggin
testing to enforce that broken-ness. Years after marking that
code obsolete, it keeps rearing its head in the unit tests.
Its like that dying dragon that whacks you with its tail on
the way out. I try to be stoic, but I can't be all the time.
I have to sometimes express frustration.