Remove GroupLink from query engine
This removes a big chunck of complex interlocking code. What seems
remarkable is that the replacement, GroupValue, is both smaller,
lines-of-code wise, more flexible (user-defined equivalence relations)
and "less integrated" -- which is kind of a good thing, here: less
integration means more freeedom to move and re-arrange, more options
and flexibility. That is, the c onnectivity was not needed, it was a
tangle, rather than being functional. Hmmm. Interesting.